+++ cmd/delete-vault.go
@@ -10,81 +10,57 @@
 // # Description: Easily push, pull and encrypt tofu and terraform statefiles from Vault. #
 // ########################################################################################
 
-package main
+package cmd
 
 import (
-	"flag"
-	"fmt"
-	"os"
-	"vaultify/cmd"
+	"bytes"
+	"log"
+	"os/exec"
+	"strings"
 )
 
-func main() {
-
-	// Define flags for version and help, allowing both -v and version, -h and help
-	var (
-		versionFlag     bool
-		versionFlagLong bool
-		helpFlag        bool
-		helpFlagLong    bool
-	)
-
-	flag.BoolVar(&versionFlag, "v", false, "Prints the version of the program")
-	flag.BoolVar(&versionFlagLong, "version", false, "Prints the version of the program")
-	flag.BoolVar(&helpFlag, "h", false, "Prints the help information")
-	flag.BoolVar(&helpFlagLong, "help", false, "Prints the help information")
-
-	flag.Parse()
-
-	switch {
-	case versionFlag || versionFlagLong:
-		cmd.Version()
-	case helpFlag || helpFlagLong:
-		cmd.Help()
-	default:
-		if len(os.Args) < 2 {
-			fmt.Println("Usage: \033[33mvaultify\033[0m [command]")
-			fmt.Println("Use \033[33m'vaultify -h'\033[0m for help.")
-			return
-		}
-
-		switch os.Args[1] {
-		case "init":
-			cmd.Init()
-		case "validate":
-			cmd.Validate()
-		case "compare":
-			cmd.Compare()
-		case "update":
-			cmd.Update()
-		case "wrap":
-			cmd.Wrap()
-		case "unwrap":
-			cmd.Unwrap()
-		case "pull":
-			cmd.Pull()
-		case "push":
-			cmd.Push()
-		case "status":
-			cmd.Status()
-		case "configure":
-			cmd.Configure()
-		case "delete":
-			cmd.Delete()
-		case "path":
-			cmd.Path()
-		case "retrieve":
-			cmd.Retrieve()
-		case "publish":
-			cmd.Publish()
-		case "permissions":
-			cmd.TokenPermissions()
-		case "install-vault":
-			cmd.InstallVault()
-		case "delete-vault":
-			cmd.DeleteVault()
-		default:
-			fmt.Printf("Unknown command: \033[33m%s\033[0m\n", os.Args[1])
-		}
+func containerExists(containerName string) bool {
+	cmd := exec.Command("docker", "ps", "-a", "-q", "-f", "name="+containerName)
+	var out bytes.Buffer
+	cmd.Stdout = &out
+	if err := cmd.Run(); err != nil {
+		log.Fatalf("Failed to query Docker containers: %v", err)
 	}
+	return strings.TrimSpace(out.String()) != ""
+}
+
+func DeleteVault() {
+	if !isDockerInstalled() {
+		log.Fatal("\033[33mDocker\033[0m is not installed.")
+	}
+
+	if !isDockerRunning() {
+		log.Fatal("\033[33mDocker\033[0m is not running. Please start the \033[33mDocker daemon\033[0m.")
+	}
+
+	const containerName = "vault-raft-backend"
+	const volumeName = "docker_vault_data"
+
+	if containerExists(containerName) {
+		if containerIsRunning(containerName) {
+			// Stop the running Vault container
+			stopCmd := exec.Command("docker", "stop", containerName)
+			if err := stopCmd.Run(); err != nil {
+				log.Fatalf("Failed to stop Vault Docker container: \033[33m%v\033[0m", err)
+			}
+		}
+
+		// Remove the Vault container
+		rmCmd := exec.Command("docker", "rm", containerName)
+		if err := rmCmd.Run(); err != nil {
+			log.Fatalf("Failed to remove Vault Docker container: \033[33m%v\033[0m", err)
+		}
+	} else {
+		log.Printf("\033[33mNo\033[0m Docker container with the name \033[33m'%s'\033[0m found. Skipping stop and remove commands.", containerName)
+	}
+
+	deleteVolume("docker_vault_data")
+	deleteVolume("vault_data")
+
+	log.Println("Vault Docker \033[33mcontainer\033[0m and \033[33mvolume\033[0m have been cleaned up.")
 }